refactoring views/controllers
