+++ Code/resources/text/sqlCommands.txt
@@ -1,0 +1,124 @@
+CREATE TABLE IF NOT EXISTS TypePaiement (
+idTypePaiement INT NOT NULL,
+nomPaiement VARCHAR(45) NOT NULL,
+PRIMARY KEY (idTypePaiement)
+);
+
+CREATE TABLE IF NOT EXISTS Contrat (
+idContrat INT NOT NULL,
+idClient INT NOT NULL,
+idVehicule INT NOT NULL,
+dateDeDepart DATETIME NOT NULL,
+dateDeRetour DATETIME NOT NULL,
+assurance BINARY NOT NULL,
+forfaitUsure BINARY NOT NULL,
+paiementCredit BINARY NOT NULL,
+PRIMARY KEY (idContrat),
+FOREIGN KEY (idClient) REFERENCES Client (idClient),
+FOREIGN KEY (idVehicule) REFERENCES Vehicule (idVehicule)
+);
+
+CREATE TABLE IF NOT EXISTS Employe (
+idEmploye INT NOT NULL,
+prenom VARCHAR(45) NOT NULL,
+nom VARCHAR(45) NOT NULL,
+telephone VARCHAR(45) NOT NULL,
+adresse VARCHAR(45) NOT NULL,
+PRIMARY KEY (idEmploye)
+);
+
+CREATE TABLE IF NOT EXISTS TypeUtilisateur (
+idTypeUtilisateur INT NOT NULL,
+nomTypeUtilisateur VARCHAR(45) NOT NULL,
+PRIMARY KEY (idTypeUtilisateur)
+);
+
+CREATE TABLE IF NOT EXISTS Utilisateur (
+idUtilisateur INT NOT NULL,
+idEmploye INT NOT NULL,
+nomUtilisateur VARCHAR(45) NOT NULL,
+motDePasse VARCHAR(45) NOT NULL,
+idTypeUtilisateur INT NOT NULL,
+PRIMARY KEY (idUtilisateur),
+FOREIGN KEY (idEmploye) REFERENCES Employe (idEmploye),
+FOREIGN KEY (idTypeUtilisateur) REFERENCES TypeUtilisateur (idTypeUtilisateur)
+);
+
+
+CREATE TABLE IF NOT EXISTS CouleurVehicule (
+idCouleurVehicule INT NOT NULL,
+nomCouleur VARCHAR(45) NOT NULL,
+PRIMARY KEY (idCouleurVehicule)
+);
+
+CREATE TABLE IF NOT EXISTS MarqueVehicule (
+idMarqueVehicule INT NOT NULL,
+nomMarque VARCHAR(45) NOT NULL,
+PRIMARY KEY (idMarqueVehicule)
+);
+
+CREATE TABLE IF NOT EXISTS TypeVehicule (
+idTypeVehicule INT NOT NULL,
+nomType VARCHAR(45) NOT NULL,
+PRIMARY KEY (idTypeVehicule)
+);
+
+CREATE TABLE IF NOT EXISTS Vehicule (
+idVehicule INT NOT NULL,
+idMarqueVehicule INT NOT NULL,
+idTypeVehicule INT NOT NULL,
+idCouleurVehicule INT NOT NULL,
+PRIMARY KEY (idVehicule),
+FOREIGN KEY (idMarqueVehicule) REFERENCES MarqueVehicule (idMarqueVehicule),
+FOREIGN KEY (idTypeVehicule) REFERENCES TypeVehicule (idTypeVehicule),
+FOREIGN KEY (idCouleurVehicule) REFERENCES CouleurVehicule (idCouleurVehicule)
+);
+
+CREATE TABLE IF NOT EXISTS TypePermis (
+idTypePermis INT NOT NULL,
+nomType VARCHAR(45) NOT NULL,
+PRIMARY KEY (idTypePermis)
+);
+
+CREATE TABLE IF NOT EXISTS Permis (
+idPermis INT NOT NULL,
+idTypePermis INT NOT NULL,
+PRIMARY KEY (idPermis),
+FOREIGN KEY (idTypePermis) REFERENCES TypePermis (idTypePermis)
+);
+
+CREATE TABLE IF NOT EXISTS Client (
+idClient INT NOT NULL,
+prenom VARCHAR(45) NOT NULL,
+nom VARCHAR(45) NOT NULL,
+dateDeNaissance DATE NOT NULL,
+telephone VARCHAR(45) NOT NULL,
+adresse VARCHAR(45) NOT NULL,
+idPermis INT NOT NULL,
+PRIMARY KEY (idClient),
+FOREIGN KEY (idPermis) REFERENCES Permis (idPermis)
+);
+
+
+CREATE TABLE IF NOT EXISTS Inspection (
+idInspection INT NOT NULL,
+idVehicule INT NOT NULL,
+idEmploye INT NOT NULL,
+dateInspection DATE NOT NULL,
+essence FLOAT NOT NULL,
+montantAccident FLOAT NOT NULL,
+PRIMARY KEY (idInspection),
+FOREIGN KEY (idVehicule) REFERENCES Vehicule (idVehicule),
+FOREIGN KEY (idEmploye) REFERENCES Employe (idEmploye)
+);
+
+
+CREATE TABLE IF NOT EXISTS Facture (
+idFacture INT NOT NULL,
+idContrat INT NOT NULL,
+idInspection INT NOT NULL,
+dateFacture DATE NOT NULL,
+PRIMARY KEY (idFacture),
+FOREIGN KEY (idContrat) REFERENCES MarqueVehicule (idMarqueVehicule),
+FOREIGN KEY (idInspection) REFERENCES TypeVehicule (idTypeVehicule)
+);
