+++ Code/src/locosys/controller/ContratLocationController.java
@@ -1,0 +1,16 @@
+package locosys.controller;
+
+import java.time.LocalDateTime;
+
+public class ContratLocationController {
+	
+	
+	public double calculateDuree(LocalDateTime dateHeureDepart, LocalDateTime dateHeureRetour) {
+		return 1;
+	}
+	
+	public double calculateMontantBase(int duree) {
+		return 1;
+	}
+
+}
