+++ Code/src/locosys/controller/EmployeGestionPageController.java
@@ -1,0 +1,5 @@
+package locosys.controller;
+
+public class EmployeGestionPageController {
+
+}
