+++ Code/src/locosys/controller/FactureController.java
@@ -1,0 +1,5 @@
+package locosys.controller;
+
+public interface FactureController {
+
+}
