+++ Code/src/locosys/controller/UtilisateurController.java
@@ -1,0 +1,5 @@
+package locosys.controller;
+
+public class UtilisateurController {
+
+}
