+++ Code/src/locosys/views/AdminPages.java
@@ -22,7 +22,6 @@ import java.awt.Component;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 
-import main.AppLogin;
 import com.toedter.calendar.JDateChooser;
 
 import locosys.controller.AppPagesController;
