+++ Code/src/main/Main.java
@@ -2,6 +2,7 @@ package main;
 
 import java.awt.EventQueue;
 import locosys.controller.AppLoginController;
+import locosys.views.AppLogin;
 import locosys.views.UserPages;
 
 public class Main {
@@ -10,13 +11,7 @@ public class Main {
 	 * Create the database.
 	 * Launch the application.
 	 */
-	public static void main(String[] args) {
-		
-
-		// Only need to be run once
-		//ConnectionBD.createNewDatabase("locosys.db");
-		//ConnectionBD.insertIntoDB();
-		
+	public static void main(String[] args) {	
 
 		EventQueue.invokeLater(new Runnable() {
 			public void run() {
